Document delete endpoint, update README, rename files to use kebab-case
